+++ src/modules/transactions/transactions.service.spec.ts
@@ -3,10 +3,11 @@ import { plainToClassFromExist } from 'class-transformer';
 import { Test, TestingModule } from '@nestjs/testing';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { ErrorTypeEnum } from 'src/common/enums';
+import { CurrencyEnum, ErrorTypeEnum } from 'src/common/enums';
 import { DatabaseModule } from 'src/database';
 import { ConfigModule } from 'src/config';
 
+import { ExchangeRateModule } from '../exchangerate';
 import { AccountEntity } from '../accounts/entities';
 import { CategoriesModule } from '../categories';
 import { AccountsModule } from '../accounts';
@@ -21,9 +22,11 @@ describe('TransactionsService', () => {
     id: 'd2727cf0-8631-48ea-98fd-29d7404b1bd2',
     amount: 10,
     description: 'Description',
+    date: new Date(),
     type: TransactionTypeEnum.EXPENSE,
     owner: {
       id: '067f2f3e-b936-4029-93d6-b2f58ae4f489',
+      baseCurrency: CurrencyEnum.UAH,
     },
     account: {
       id: 'd2727cf0-8631-48ea-98fd-29d7404b1bca',
@@ -43,6 +46,7 @@ describe('TransactionsService', () => {
         DatabaseModule,
         AccountsModule,
         CategoriesModule,
+        ExchangeRateModule,
       ],
       providers: [TransactionsService],
     }).compile();
@@ -56,14 +60,14 @@ describe('TransactionsService', () => {
 
   describe('createOne', () => {
     it('should be return transaction entity', async () => {
-      const received = await service.createOne(expected);
+      const received = await service.createOne(expected, expected.owner);
       expect(received).toBeInstanceOf(TransactionEntity);
       expect(received.id).toEqual(expected.id);
     });
 
     it('should be return conflict exception', async () => {
       const error = new ConflictException(ErrorTypeEnum.TRANSACTION_ALREADY_EXIST);
-      return service.createOne(expected).catch((err) => {
+      return service.createOne(expected, expected.owner).catch((err) => {
         expect(err).toBeInstanceOf(ConflictException);
         expect(err).toEqual(error);
       });
@@ -73,7 +77,7 @@ describe('TransactionsService', () => {
   describe('selectAll', () => {
     it('should be return transactions pagination entity', async () => {
       const received = await service.selectAll();
-      expect(received).toHaveLength(expect.any(Number));
+      expect(received.length).toEqual(expect.any(Number));
     });
 
     it('should be return not found exception', async () => {
@@ -104,10 +108,9 @@ describe('TransactionsService', () => {
 
   describe('updateOne', () => {
     it('should be return transaction entity', async () => {
-      const received = await service.updateOne(
-        { id: expected.id },
-        { description: 'Description 2' },
-      );
+      const received = await service.updateOne({ id: expected.id }, expected.owner, {
+        description: 'Description 2',
+      });
       expect(received).toBeInstanceOf(TransactionEntity);
       expect(received.description).not.toEqual(expected.description);
       expect(received.id).toEqual(expected.id);
@@ -115,23 +118,8 @@ describe('TransactionsService', () => {
 
     it('should be return conflict exception', async () => {
       const error = new NotFoundException(ErrorTypeEnum.TRANSACTION_NOT_FOUND);
-      return service.updateOne({ id: expected.id }, {}).catch((err) => {
+      return service.updateOne({ id: expected.id }, expected.owner, {}).catch((err) => {
         expect(err).toBeInstanceOf(NotFoundException);
-        expect(err).toEqual(error);
-      });
-    });
-
-    it('should be return not found exception', async () => {
-      const error = new ConflictException(ErrorTypeEnum.TRANSACTION_ALREADY_EXIST);
-
-      jest
-        .spyOn(service, 'selectOne')
-        .mockImplementationOnce(async () => ({ id: '' } as TransactionEntity));
-      jest.spyOn(service, 'genReverseState').mockImplementationOnce(() => new TransactionEntity());
-      jest.spyOn(service, 'processOne').mockImplementationOnce(async () => new TransactionEntity());
-
-      return service.updateOne({ id: '' }, { description: null }).catch((err) => {
-        expect(err).toBeInstanceOf(ConflictException);
         expect(err).toEqual(error);
       });
     });
@@ -139,7 +127,7 @@ describe('TransactionsService', () => {
 
   describe('deleteOne', () => {
     it('should be return transaction entity', async () => {
-      const received = await service.deleteOne({ id: expected.id });
+      const received = await service.deleteOne({ id: expected.id }, expected.owner);
       expect(received).toBeInstanceOf(TransactionEntity);
     });
 
@@ -150,7 +138,7 @@ describe('TransactionsService', () => {
       jest.spyOn(service, 'genReverseState').mockImplementationOnce(() => new TransactionEntity());
       jest.spyOn(service, 'processOne').mockImplementationOnce(async () => new TransactionEntity());
 
-      return service.deleteOne({ id: '' }).catch((err) => {
+      return service.deleteOne({ id: '' }, expected.owner).catch((err) => {
         expect(err).toBeInstanceOf(NotFoundException);
         expect(err).toEqual(error);
       });
